moved public interface tests to integration test folder
